refactor : Type 클래스, pieces 패키지로 이동
  > Type 클래스 위치 변경
    - chess 패키지 → pieces 패키지 이동
