+++ src/main/java/net/woody/pieces/Type.java
@@ -1,4 +1,4 @@
-package net.woody.chess;
+package net.woody.pieces;
 
 public enum Type {
